Add error messaging for when DB is exhausted
